Write some test case for Heroku
